Updated facades and added exceptions

+++ BusinessLogic/LoggedInAirlineFacade.cs
@@ -9,54 +9,85 @@ namespace BusinessLogic
     {
         public void CancelFlight(LoginToken<AirlineCompany> token, Flight flight)
         {
-            _flightDAO.Remove(flight);
+            if (token != null)
+                _flightDAO.Remove(flight);
+            else
+                throw new WasntActivatedByAirlineException();
         }
 
         public void ChangeMyPassword(LoginToken<AirlineCompany> token, string oldPassword, string newPassword)
         {
-            token.User.User.Password = newPassword;
-            _userDAO.Update(token.User.User);
+            if (token != null)
+            {
+                if (token.User.User.Password != oldPassword)
+                    throw new WrongCredentialsException();
+                token.User.User.Password = newPassword;
+                _userDAO.Update(token.User.User);
+            }
+            else
+                throw new WasntActivatedByAirlineException();
         }
 
         public void CreateFlight(LoginToken<AirlineCompany> token, Flight flight)
         {
-            flight.Airline_Company = token.User;
-            flight.Airline_Company_Id = token.User.Id;
-            _flightDAO.Add(flight);
+            if (token != null)
+            {
+                flight.Airline_Company = token.User;
+                flight.Airline_Company_Id = token.User.Id;
+                _flightDAO.Add(flight);
+            }
+            else
+                throw new WasntActivatedByAirlineException();
         }
 
         public IList<Flight> GetAllFlights(LoginToken<AirlineCompany> token)
         {
-            List<Flight> flights = new List<Flight>();
-            foreach (Flight flight in _flightDAO.GetAll())
+            if (token != null)
             {
-                if (flight.Airline_Company == token.User)
-                    flights.Add(flight);
+                List<Flight> flights = new List<Flight>();
+                foreach (Flight flight in _flightDAO.GetAll())
+                {
+                    if (flight.Airline_Company == token.User)
+                        flights.Add(flight);
+                }
+                return flights;
             }
-            return flights;
+            else
+                throw new WasntActivatedByAirlineException();
         }
 
         public IList<Ticket> GetAllTickets(LoginToken<AirlineCompany> token)
         {
-            List<Ticket> tickets = new List<Ticket>();
-            IList<Flight> flights = GetAllFlights(token);
-            IList<Ticket> tickets_list = _ticketDAO.GetAll();
-            foreach (Ticket ticket in tickets_list)
+            if (token != null)
             {
-                if (flights.Contains(ticket.Flight))
-                    tickets.Add(ticket);
+                List<Ticket> tickets = new List<Ticket>();
+                IList<Flight> flights = GetAllFlights(token);
+                IList<Ticket> tickets_list = _ticketDAO.GetAll();
+                foreach (Ticket ticket in tickets_list)
+                {
+                    if (flights.Contains(ticket.Flight))
+                        tickets.Add(ticket);
+                }
+                return tickets;
             }
-            return tickets;
+            else
+                throw new WasntActivatedByAirlineException();
         }
 
         public void MofidyAirlineDetails(LoginToken<AirlineCompany> token, AirlineCompany airline)
         {
-            _airlineDAO.Update(airline);
+            if (token != null)
+                _airlineDAO.Update(airline);
+            else
+                throw new WasntActivatedByAirlineException();
         }
 
         public void UpdateFlight(LoginToken<AirlineCompany> token, Flight flight)
         {
-            _flightDAO.Update(flight);
+            if (token != null)
+                _flightDAO.Update(flight);
+            else
+                throw new WasntActivatedByAirlineException();
         }
     }
 }
